Added the complete codebase
Added the Scrapy spider code for extracting contact information from companies' websites based on their names. The spider uses web scraping techniques to find the official websites of the companies and then scrapes the contact information from their 'Contact Us' pages.

The spider script is written in Python and utilizes various libraries, including Scrapy, pandas, phonenumbers, re, requests, BeautifulSoup (bs4), email_validator, fake_useragent, googlesearch-python, and selenium. Additionally, the spider requires the 'chromedriver' executable to interact with the Chrome browser for certain tasks.

The README.md file is included with detailed instructions on how to set up the project. It explains the dependencies required, external programs needed, and provides guidance on how to run the spider using the 'scrapy crawl example' command.

This marks the beginning of the project, and further enhancements and improvements will be made in subsequent commits.